Refactor functions in seperate module
